Separate out DigitalOcean plugin from Admin
